requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
PyJWT==2.3.0
werkzeug==2.0.3  # Add this line
